requirements.txt
abydos==0.5.0
affine==2.3.0
alabaster==0.7.12
apparmor==3.0.3
appdirs==1.4.4
application-utility==1.3.2
argh==0.26.2
argon2-cffi==20.1.0
arrow==1.1.1
astroid==2.7.3
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
Beaker==1.11.0
binaryornot==0.4.4
black==21.8b0
bleach==4.1.0
btrfsutil==5.13.1
CacheControl==0.12.6
cached-property==1.5.2
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==1.6.0
colorama==0.4.4
contextlib2==0.6.0.post1
cookiecutter==1.7.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.3
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
distro==1.6.0
dnspython==1.16.0
docker==5.0.2
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
flake8==3.9.2
future==0.18.2
gmpy2==2.1.0rc1
helpdev==0.7.1
html5lib==1.1
idna==3.2
imagesize==1.2.0
importlib-metadata==4.8.1
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.0.3
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.3
jedi==0.18.0
jeepney==0.7.1
jellyfish==0.8.8
Jinja2==3.0.1
jinja2-time==0.2.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keyring==23.0.1
keyutils==0.6
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
LibAppArmor==3.0.3
libtorrent===1.2.14-build-libtorrent-rasterbar-src-libtorrent-rasterbar-1.2.14-bindings-python
lit==12.0.1.dev0
louis==3.19.0
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
meson==0.59.1
mistune==0.8.4
more-itertools==8.8.0
mpmath==1.2.1
msgpack==1.0.2
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netsnmp-python==1.0a1
notebook==6.4.3
npyscreen==4.10.5
numpy==1.20.3
numpydoc==1.1.0
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.21.5
pandas==1.3.2
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
pathspec==0.9.0
pathtools==0.1.2
pep517==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
planet==1.4.7
platformdirs==2.0.2
pluggy==0.13.1
ply==3.11
poyo==0.5.0
progress==1.5
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psutil==5.8.0
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.7.0
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.10.0
PyGObject==3.40.1
pylint==2.10.2
pyls-black==0.4.6
pyls-spyder==0.3.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.4
pyrsistent==0.18.0
pysmbc==1.0.23
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-Levenshtein==0.12.2
python-slugify==5.0.2
pytz==2021.1
pyxDamerauLevenshtein==1.7.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.2.1
QDarkStyle==3.0.2
qstylizer==0.2.0
QtAwesome==1.0.2
qtconsole==5.1.1
QtPy==1.9.0
rados==2.0.0
rasterio==1.2.8
rbd==2.0.0
regex==2021.7.6
reportlab==3.6.1
requests==2.26.0
requests-futures==0.9.7
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
rope==0.19.0
scipy==1.7.1
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
snowballstemmer==2.1.0
snuggs==1.4.7
sortedcontainers==2.4.0
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.0.1
spyder-kernels==2.0.1
sympy==1.8
team==1.0
terminado==0.11.1
testpath==0.4.4
text-unidecode==1.3
textdistance==4.2.1
texttable==1.6.4
three-merge==0.1.1
tinycss2==1.1.0
toml==0.10.2
tomli==1.2.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.2
udiskie==2.3.3
ujson==4.1.0
urllib3==1.26.6
watchdog==0.10.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.1.1
yapf==0.31.0
zipp==3.5.0
